[ADD] finding content component , adap form accordingly
